requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
python-dotenv==1.1.1
pydantic==2.11.7
google-generativeai==0.8.5
soundfile==0.13.1
numpy==2.2.6
torch
torchaudio
nemo_toolkit[asr]
huggingface-hub>=0.34.0,<1.0
transformers>=4.50,<5